Add basic structure for the GameInfo rewrite
Basically, there's a static class which holds the properties to the needed data
Then there's a bunch of threads taking care of filling of those properties.
This runs forever and only shuts when the application will be closed.

The filling requires the _correct_ offsets and a reference to the SC2 process.
Using that, there's a dedicated thread for each needed variable making the
filling more independent and split between cores. Thus: More efficient for one
property.

Not sure about the overhead though.

The reason for the rewrite is kind of logical: There is only one instance that
grabs the information from SC2. Currently, I pass the references to those
variables around - a huge overhead and bad design at it's core.
A static class exists only one time per application. So there's a huge
convenient reason; logic.

The drawback is the lack of plugin support. You have to send those data to the
plugin by hand.